chore: 🤖 review changes
✅ Closes: #482
