--- apps/web/src/app/(protectedPage)/(mainPage)/forms/loading.tsx
@@ -1,29 +1,0 @@
-import { Skeleton } from "@convoform/ui";
-
-export default function Loading() {
-  return (
-    <div className="pb-5 container px-4 ms-0 space-y-6">
-      <div className="space-y-4">
-        <div className="flex items-baseline justify-between gap-3">
-          <div className="grow flex items-center gap-2">
-            <Skeleton className="h-8 w-8" />
-            <Skeleton className="h-8 w-24" />
-          </div>
-        </div>
-        <Skeleton className="h-9 w-48" />
-      </div>
-      <div className="grid grid-cols-1 gap-6 sm:grid-cols-2 lg:grid-cols-3">
-        {[...Array(6)].map((_, i) => (
-          // biome-ignore lint/suspicious/noArrayIndexKey: Static skeleton list
-          <div key={i} className="flex flex-col space-y-3">
-            <Skeleton className="h-[200px] w-full rounded-xl" />
-            <div className="space-y-2">
-              <Skeleton className="h-4 w-full" />
-              <Skeleton className="h-4 w-[80%]" />
-            </div>
-          </div>
-        ))}
-      </div>
-    </div>
-  );
-}
