+++ packages/api/.env.test
@@ -1,3 +1,4 @@
 NEXT_PUBLIC_APP_URL=http://localhost:3000
 INTERNAL_EMAIL_API_SECRET=test
 CLERK_SECRET_KEY=test
+ENCRYPTION_KEY=test
