+++ packages/api/package.json
@@ -20,7 +20,7 @@
     "@convoform/file-storage": "workspace:*",
     "@convoform/integration": "workspace:*",
     "@convoform/rate-limiter": "workspace:*",
-    "drizzle-orm": "0.45.1",
+    "drizzle-orm": "catalog:",
     "@t3-oss/env-core": "0.13.10",
     "@trpc/client": "11.0.0-rc.571",
     "@trpc/server": "11.0.0-rc.571",
