+++ packages/common/src/crypto.test.ts
@@ -1,4 +1,4 @@
-import { describe, expect, it } from "vitest";
+import { describe, expect, it } from "bun:test";
 import { decrypt, encrypt, signPayload, verifySignature } from "./crypto";
 
 describe("Crypto Utils", () => {
