+++ packages/db/package.json
@@ -24,7 +24,7 @@
     "@neondatabase/serverless": "1.0.2",
     "@paralleldrive/cuid2": "2.2.2",
     "@t3-oss/env-core": "0.13.10",
-    "drizzle-orm": "0.45.1",
+    "drizzle-orm": "catalog:",
     "drizzle-zod": "0.8.3",
     "zod": "catalog:"
   },
