+++ turbo.json
@@ -40,7 +40,8 @@
     "AXIOM_OTLP_TOKEN",
     "EMAIL_FROM",
     "RESEND_API_KEY",
-    "INTERNAL_EMAIL_API_SECRET"
+    "INTERNAL_EMAIL_API_SECRET",
+    "ENCRYPTION_KEY"
   ],
   "globalDependencies": [
     ".env"
